+++ vsf/interfaces/cpu/stm32/gpio/STM32_GPIO.h
@@ -16,7 +16,7 @@
 
 vsf_err_t stm32_gpio_init(uint8_t index);
 vsf_err_t stm32_gpio_fini(uint8_t index);
-vsf_err_t stm32_gpio_config_pin(uint8_t index, uint8_t pin_idx, uint8_t mode);
+vsf_err_t stm32_gpio_config_pin(uint8_t index, uint8_t pin_idx, uint32_t mode);
 vsf_err_t stm32_gpio_config(uint8_t index, uint32_t pin_mask, uint32_t io, 
 							uint32_t pull_en_mask, uint32_t input_pull_mask);
 vsf_err_t stm32_gpio_set(uint8_t index, uint32_t pin_mask);
